+++ src/main/java/com/example/webgame/service/TypeService.java
@@ -5,5 +5,5 @@ import java.util.List;
 public interface TypeService {
     public List<String> getAllTypeName();
 
-    public String getTypeName(Long type_id);
+    public String getTypeName(Integer type_id);
 }
